switched to using a PCL instead of the netstandard library
