+++ HalClient.Net/Parser/LinkObject.cs
@@ -15,7 +15,7 @@ namespace HalClient.Net.Parser
 		public Uri Profile { get; private set; }
 		public string Title { get; set; }
 		public string HrefLang { get; set; }
-		
+
 		public ILinkObject ResolveTemplated(Func<UriTemplate, string> hrefResolver)
 		{
 			if (!Templated)
