+++ HalClient.Net/Properties/AssemblyInfo.cs
@@ -2,11 +2,5 @@
 using System.Runtime.InteropServices;
 
 [assembly: AssemblyTitle("HalClient.Net")]
-[assembly: AssemblyDescription("Specialized http client for communicating with API's that support the HAL media type (https://github.com/wis3guy/HalClient.Net)")]
-[assembly: AssemblyCompany("Geoffrey Braaf, WIS3GUY.NET")]
-[assembly: AssemblyProduct("HalClient.Net")]
-[assembly: AssemblyCopyright("Copyright (C) Geoffrey Braaf, WIS3GUY.NET")]
 [assembly: ComVisible(false)]
 [assembly: Guid("cca82297-341f-4b5f-a0fe-e650b25b72eb")]
-[assembly: AssemblyVersion("2.0.0.*")]
-[assembly: AssemblyInformationalVersion("2.0.0")]
