+++ HalClient.Net/Properties/SharedAssemblyInfo.cs
@@ -1,0 +1,8 @@
+﻿using System.Reflection;
+
+[assembly: AssemblyDescription("Specialized http client for communicating with API's that support the HAL media type (https://github.com/wis3guy/HalClient.Net)")]
+[assembly: AssemblyCompany("Geoffrey Braaf, WIS3GUY.NET")]
+[assembly: AssemblyProduct("HalClient.Net")]
+[assembly: AssemblyCopyright("Copyright (C) Geoffrey Braaf, WIS3GUY.NET")]
+[assembly: AssemblyVersion("2.0.0.*")]
+[assembly: AssemblyInformationalVersion("2.0.0")]
