+++ HalClient.Net/packages.config
@@ -1,6 +1,6 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="Microsoft.AspNet.WebApi.Client" version="5.2.3" targetFramework="net45" />
-  <package id="Newtonsoft.Json" version="9.0.1" targetFramework="net451" />
-  <package id="Tavis.UriTemplates" version="1.1.0" targetFramework="net451" />
+  <package id="Newtonsoft.Json" version="10.0.2" targetFramework="net451" />
+  <package id="System.Net.Http" version="4.3.2" targetFramework="net451" />
+  <package id="Tavis.UriTemplates" version="1.1.1" targetFramework="net451" />
 </packages>
