--- HalClient.NetStandard/CachingBehavior.cs
@@ -1,9 +1,0 @@
-namespace HalClient.Net
-{
-	public enum CachingBehavior
-	{
-		Never,
-		PerClient,
-		Once
-	}
-}
