--- HalClient.NetStandard/HalClient.NetStandard.csproj
@@ -1,14 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>netstandard1.4</TargetFramework>
-    <RootNamespace>HalClient.Net</RootNamespace>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Newtonsoft.Json" Version="10.0.2" />
-    <PackageReference Include="System.Net.Http" Version="4.3.2" />
-    <PackageReference Include="Tavis.UriTemplates" Version="1.1.1" />
-  </ItemGroup>
-
-</Project>
