--- HalClient.NetStandard/IHalHttpResponseMessage.cs
@@ -1,12 +1,0 @@
-using System.Net.Http;
-using HalClient.Net.Parser;
-
-namespace HalClient.Net
-{
-	public interface IHalHttpResponseMessage
-	{
-		HttpResponseMessage Message { get; }
-		bool IsHalResponse { get; }
-		IRootResourceObject Resource { get; }
-	}
-}
