--- HalClient.NetStandard/MediaType.cs
@@ -1,7 +1,0 @@
-namespace HalClient.Net
-{
-	internal static class MediaType
-	{
-		public const string ApplicationHalPlusJson = "application/hal+json";
-	}
-}
