--- HalClient.NetStandard/Parser/IEmbeddedResourceObject.cs
@@ -1,6 +1,0 @@
-namespace HalClient.Net.Parser
-{
-	public interface IEmbeddedResourceObject : IResourceObject, IHaveLinkRelation
-	{
-	}
-}
