--- HalClient.NetStandard/Parser/IHalJsonParser.cs
@@ -1,7 +1,0 @@
-namespace HalClient.Net.Parser
-{
-	public interface IHalJsonParser
-	{
-		HalJsonParseResult Parse(string json);
-	}
-}
