--- HalClient.NetStandard/Parser/IHaveLinkRelation.cs
@@ -1,7 +1,0 @@
-namespace HalClient.Net.Parser
-{
-	public interface IHaveLinkRelation
-	{
-		string Rel { get; }
-	}
-}
