--- HalClient.NetStandard/Parser/IRootResourceObject.cs
@@ -1,9 +1,0 @@
-using System;
-
-namespace HalClient.Net.Parser
-{
-	public interface IRootResourceObject : IResourceObject
-	{
-		Uri GetDocumentationUri(IHaveLinkRelation link);
-	}
-}
