--- HalClient.NetStandard/Parser/IStateValue.cs
@@ -1,9 +1,0 @@
-namespace HalClient.Net.Parser
-{
-	public interface IStateValue
-	{
-		string Name { get; }
-		string Value { get; }
-		string Type { get; }
-	}
-}
