+++ HalClient.Portable/HalClient.Portable.csproj
@@ -1,0 +1,129 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<Project ToolsVersion="15.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
+  <Import Project="$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props" Condition="Exists('$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props')" />
+  <PropertyGroup>
+    <MinimumVisualStudioVersion>14.0</MinimumVisualStudioVersion>
+    <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
+    <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
+    <ProjectGuid>{FBB9507B-1262-4C42-80C8-83E0356582B5}</ProjectGuid>
+    <OutputType>Library</OutputType>
+    <RootNamespace>HalClient</RootNamespace>
+    <AssemblyName>HalClient.Portable</AssemblyName>
+    <DefaultLanguage>en-US</DefaultLanguage>
+    <FileAlignment>512</FileAlignment>
+    <ProjectTypeGuids>{786C830F-07A1-408B-BD7F-6EE04809D6DB};{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}</ProjectTypeGuids>
+    <TargetFrameworkProfile>
+    </TargetFrameworkProfile>
+    <TargetFrameworkVersion>v5.0</TargetFrameworkVersion>
+  </PropertyGroup>
+  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Debug|AnyCPU' ">
+    <DebugSymbols>true</DebugSymbols>
+    <DebugType>full</DebugType>
+    <Optimize>false</Optimize>
+    <OutputPath>bin\Debug\</OutputPath>
+    <DefineConstants>DEBUG;TRACE</DefineConstants>
+    <ErrorReport>prompt</ErrorReport>
+    <WarningLevel>4</WarningLevel>
+  </PropertyGroup>
+  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Release|AnyCPU' ">
+    <DebugType>pdbonly</DebugType>
+    <Optimize>true</Optimize>
+    <OutputPath>bin\Release\</OutputPath>
+    <DefineConstants>TRACE</DefineConstants>
+    <ErrorReport>prompt</ErrorReport>
+    <WarningLevel>4</WarningLevel>
+  </PropertyGroup>
+  <ItemGroup>
+    <Compile Include="..\HalClient.Net\CachingBehavior.cs">
+      <Link>CachingBehavior.cs</Link>
+    </Compile>
+    <Compile Include="..\HalClient.Net\HalHttpClient.cs">
+      <Link>HalHttpClient.cs</Link>
+    </Compile>
+    <Compile Include="..\HalClient.Net\HalHttpClientConfiguration.cs">
+      <Link>HalHttpClientConfiguration.cs</Link>
+    </Compile>
+    <Compile Include="..\HalClient.Net\HalHttpClientFactory.cs">
+      <Link>HalHttpClientFactory.cs</Link>
+    </Compile>
+    <Compile Include="..\HalClient.Net\HalHttpClientFactoryBase.cs">
+      <Link>HalHttpClientFactoryBase.cs</Link>
+    </Compile>
+    <Compile Include="..\HalClient.Net\HalHttpRequestException.cs">
+      <Link>HalHttpRequestException.cs</Link>
+    </Compile>
+    <Compile Include="..\HalClient.Net\HalHttpResponseMessage.cs">
+      <Link>HalHttpResponseMessage.cs</Link>
+    </Compile>
+    <Compile Include="..\HalClient.Net\IHalHttpClient.cs">
+      <Link>IHalHttpClient.cs</Link>
+    </Compile>
+    <Compile Include="..\HalClient.Net\IHalHttpClientConfiguration.cs">
+      <Link>IHalHttpClientConfiguration.cs</Link>
+    </Compile>
+    <Compile Include="..\HalClient.Net\IHalHttpClientFactory.cs">
+      <Link>IHalHttpClientFactory.cs</Link>
+    </Compile>
+    <Compile Include="..\HalClient.Net\IHalHttpResponseMessage.cs">
+      <Link>IHalHttpResponseMessage.cs</Link>
+    </Compile>
+    <Compile Include="..\HalClient.Net\MediaType.cs">
+      <Link>MediaType.cs</Link>
+    </Compile>
+    <Compile Include="..\HalClient.Net\Parser\EmbeddedResourceObject.cs">
+      <Link>EmbeddedResourceObject.cs</Link>
+    </Compile>
+    <Compile Include="..\HalClient.Net\Parser\HalJsonParser.cs">
+      <Link>HalJsonParser.cs</Link>
+    </Compile>
+    <Compile Include="..\HalClient.Net\Parser\HalJsonParseResult.cs">
+      <Link>HalJsonParseResult.cs</Link>
+    </Compile>
+    <Compile Include="..\HalClient.Net\Parser\IEmbeddedResourceObject.cs">
+      <Link>IEmbeddedResourceObject.cs</Link>
+    </Compile>
+    <Compile Include="..\HalClient.Net\Parser\IHalJsonParser.cs">
+      <Link>IHalJsonParser.cs</Link>
+    </Compile>
+    <Compile Include="..\HalClient.Net\Parser\IHaveLinkRelation.cs">
+      <Link>IHaveLinkRelation.cs</Link>
+    </Compile>
+    <Compile Include="..\HalClient.Net\Parser\ILinkObject.cs">
+      <Link>ILinkObject.cs</Link>
+    </Compile>
+    <Compile Include="..\HalClient.Net\Parser\IResourceObject.cs">
+      <Link>IResourceObject.cs</Link>
+    </Compile>
+    <Compile Include="..\HalClient.Net\Parser\IRootResourceObject.cs">
+      <Link>IRootResourceObject.cs</Link>
+    </Compile>
+    <Compile Include="..\HalClient.Net\Parser\IStateValue.cs">
+      <Link>IStateValue.cs</Link>
+    </Compile>
+    <Compile Include="..\HalClient.Net\Parser\LinkObject.cs">
+      <Link>LinkObject.cs</Link>
+    </Compile>
+    <Compile Include="..\HalClient.Net\Parser\ResourceObjectBase.cs">
+      <Link>ResourceObjectBase.cs</Link>
+    </Compile>
+    <Compile Include="..\HalClient.Net\Parser\RootResourceObject.cs">
+      <Link>RootResourceObject.cs</Link>
+    </Compile>
+    <Compile Include="..\HalClient.Net\Parser\StateValue.cs">
+      <Link>StateValue.cs</Link>
+    </Compile>
+    <Compile Include="..\HalClient.Net\Properties\SharedAssemblyInfo.cs">
+      <Link>Properties\SharedAssemblyInfo.cs</Link>
+    </Compile>
+    <Compile Include="Properties\AssemblyInfo.cs" />
+  </ItemGroup>
+  <ItemGroup>
+    <None Include="project.json" />
+  </ItemGroup>
+  <ItemGroup>
+    <Reference Include="System">
+      <HintPath>..\..\..\Program Files (x86)\Reference Assemblies\Microsoft\Framework\.NETFramework\v4.6.1\System.dll</HintPath>
+    </Reference>
+  </ItemGroup>
+  <Import Project="$(MSBuildExtensionsPath32)\Microsoft\Portable\$(TargetFrameworkVersion)\Microsoft.Portable.CSharp.targets" />
+</Project>
